requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.4
Flask-JWT-Extended==4.5.2
pymongo==4.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
eventlet==0.33.3
python-socketio==5.9.0 